refactor in utils, jest tests, workflow updated
