+++ package.json
@@ -7,7 +7,8 @@
     "ios": "react-native run-ios",
     "lint": "eslint .",
     "start": "react-native start",
-    "pre-commit": "yarn lint && yarn tsc --noEmit"
+    "pre-commit": "yarn lint && yarn tsc --noEmit && yarn test",
+    "test": "jest --verbose"
   },
   "dependencies": {
     "@hookform/resolvers": "^3.9.1",
@@ -49,6 +50,7 @@
     "@react-native/metro-config": "0.76.2",
     "@react-native/typescript-config": "0.76.2",
     "@tanstack/eslint-plugin-query": "^5.62.1",
+    "@types/jest": "^29.5.14",
     "@types/react": "^18.2.6",
     "@types/react-test-renderer": "^18.0.0",
     "babel-jest": "^29.6.3",
