+++ src/screens/app/HomeScreen/components/MealsCalBudget.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import {Meal} from '@domain';
-import {macrosCalculations} from '@utils';
+import {calcCalsByMealType, calcMealTotals} from '@utils';
 
 import {Box, CalorieRing, Surface, Text} from '@components';
 
@@ -20,14 +20,14 @@ export function MealsCalBudget({meals}: MealsCalTableProps) {
       </Box>
       <CalorieRing
         currentCalories={
-          macrosCalculations.calculateMealTotals(meals).totalCalories
+          calcMealTotals.calculateMealTotals(meals).totalCalories
         }
         goalCalories={2000}
       />
       <Box position={'absolute'} top={60} left={20} alignItems={'center'}>
         <Text preset={'paragraphSmall'}>Breakfast</Text>
         <Text preset={'headingMedium'} color={'bluePrimary'}>
-          {macrosCalculations
+          {calcCalsByMealType
             .calculateCaloriesByMealType(meals)
             .breakfast.toFixed(0)}
         </Text>
@@ -35,7 +35,7 @@ export function MealsCalBudget({meals}: MealsCalTableProps) {
       <Box position={'absolute'} top={130} left={20} alignItems={'center'}>
         <Text preset={'paragraphSmall'}>Lunch</Text>
         <Text preset={'headingMedium'} color={'bluePrimary'}>
-          {macrosCalculations
+          {calcCalsByMealType
             .calculateCaloriesByMealType(meals)
             .lunch.toFixed(0)}
         </Text>
@@ -43,7 +43,7 @@ export function MealsCalBudget({meals}: MealsCalTableProps) {
       <Box position={'absolute'} top={130} right={20} alignItems={'center'}>
         <Text preset={'paragraphSmall'}>Snack</Text>
         <Text preset={'headingMedium'} color={'bluePrimary'}>
-          {macrosCalculations
+          {calcCalsByMealType
             .calculateCaloriesByMealType(meals)
             .snack.toFixed(0)}
         </Text>
@@ -51,7 +51,7 @@ export function MealsCalBudget({meals}: MealsCalTableProps) {
       <Box position={'absolute'} top={60} right={20} alignItems={'center'}>
         <Text preset={'paragraphSmall'}>Dinner</Text>
         <Text preset={'headingMedium'} color={'bluePrimary'}>
-          {macrosCalculations
+          {calcCalsByMealType
             .calculateCaloriesByMealType(meals)
             .dinner.toFixed(0)}
         </Text>
