+++ src/utils/index.ts
@@ -1,1 +1,6 @@
 export * from './macrosCalculations';
+export * from './calculateCaloriesByMealType';
+export * from './calculateMealTotals';
+export * from './recipeTotals';
+export * from './calculateMealItemTotals';
+export * from './calculateRecipeMacros';
