+++ src/utils/utilsMocks.ts
@@ -1,0 +1,21 @@
+const food = {
+  id: 1,
+  createdAt: 'string',
+  userId: 'string',
+  label: 'string',
+  protein: 10,
+  carbs: 10,
+  fat: 10,
+  calories: 10,
+  fibre: 10,
+  sodium: 10,
+  servSize: 10,
+  servUnit: 'string',
+  foodImg: 'string',
+  categoryId: 10,
+  isArchived: false,
+};
+
+export const utilsMocks = {
+  food,
+};
